+++ src/gfx/gfx.h
@@ -17,6 +17,6 @@ void gfx_end();
 void gfx_flush();
 
 void gfx_background(float r, float g, float b);
-void gfx_sprite2d(float x, float y, float width, float height, TextureRect texture_rect);
+void gfx_sprite2d(int x, int y, int width, int height, TextureRect texture_rect);
 
 #endif
